requirements.txt
aiohttp==3.7.4
Django==4.0.7
django-ratelimit==2.0.0
Markdown==3.1.1
Pillow
python-memcached==1.59
requests==2.22.0
beautifulsoup4==4.8.2
django-admin-interface==0.20.0
psycopg2
django-colorfield
django-registration-redux
python-dotenv